requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==23.12.1
certifi==2023.11.17
click==8.1.7
Django==5.0.1
django-multiupload==0.6.1
exceptiongroup==1.2.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
pycairo==1.25.1
PyGObject==3.46.0
PySocks==1.7.1
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
tomli==2.0.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
wsproto==1.2.0
